third_party/WebKit/LayoutTests/compositing/overflow/do-not-opt-in-with-out-of-flow-descendant-expected.txt

Iteration 0: Passed
Iteration 0, container is composited
Iteration 1: Passed
Iteration 1, container is not composited
Iteration 2: Passed
Iteration 2, container is composited
Iteration 3: Passed
Iteration 3, container is not composited
Iteration 4: Passed
Iteration 4, container is composited
Iteration 5: Passed
Iteration 5, container is not composited
Iteration 6: Passed
Iteration 6, container is composited
